app/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.9.1
openai==1.54.4
langchain==0.3.7
langchain-openai==0.2.8
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17
httpx==0.27.2
